+++ models/comments.js
@@ -1,7 +1,7 @@
 module.exports = (sequelize, DataTypes) => {
-    return sequelize.define('tags', {
-        title: {
-            type: DataTypes.STRING(30),
+    return sequelize.define('comments', {
+        contents: {
+            type: DataTypes.STRING,
             allowNULL: false,
             unique: false,
         },
